Add files via upload
This Project just have a Login and Sign-in pages which are all connected to the MySQL Database. If you want to make it work do change the Database name and Password in the sittings.py and the views.py, further your database should be having the following code in MySQL.
CREATE TABLE Users (
  email VARCHAR(255) PRIMARY KEY,
  password VARCHAR(255) NOT NULL,
  address1 VARCHAR(255) NOT NULL,
  address2 VARCHAR(255),
  city VARCHAR(255) NOT NULL,
  state VARCHAR(255) NOT NULL,
  zip VARCHAR(10) NOT NULL
);


If there are errors make sure to check the user=='root' even though its a little different(IF IT WON'T WORK).